+++ linking-platform/smart-card/src/__tests__/visual-regression/embed-card.ts
@@ -4,9 +4,11 @@ describe('Embed Card', () => {
   it('shows unresolved views', async () => {
     const url = getURL('vr-embed-card-unresolved-views');
     const page = await setup(url);
-    await page.waitForSelector('[data-testid="embed-card-forbidden-view"] img');
+    await page.waitForSelector(
+      '[data-testid="embed-card-forbidden-view-unresolved-image"]',
+    );
 
-    const image = await takeSnapshot(page, 1600);
+    const image = await takeSnapshot(page, 2050);
     expect(image).toMatchProdImageSnapshot();
   });
 });