+++ linking-platform/smart-card/src/view/EmbedCard/types.ts
@@ -31,6 +31,7 @@ export interface WithShowControlMethodProp {
 
 export interface ContextViewModel {
   icon?: ReactNode;
+  image?: string;
   text: string;
 }