+++ linking-platform/smart-card/src/view/EmbedCard/views/UnauthorisedView.tsx
@@ -1,9 +1,12 @@
 import React, { FC } from 'react';
+import { FormattedMessage } from 'react-intl-next';
 import { EmbedCardUnresolvedView } from './UnresolvedView';
-import { CelebrationImage } from '../../BlockCard/utils/constants';
+import { UnauthorisedImage } from '../constants';
 import { ExpandedFrame } from '../components/ExpandedFrame';
 import { ImageIcon } from '../components/ImageIcon';
 import { ContextViewModel } from '../types';
+import { messages } from '../../../messages';
+import { CONTENT_URL_SECURITY_AND_PERMISSIONS } from '../../../constants';
 
 export interface EmbedCardUnauthorisedViewProps {
   context?: ContextViewModel;
@@ -42,18 +45,32 @@ export const EmbedCardUnauthorisedView: FC<EmbedCardUnauthorisedViewProps> = ({
       testId={testId}
     >
       <EmbedCardUnresolvedView
-        image={CelebrationImage}
+        image={context?.image ?? UnauthorisedImage}
         title="connect_link_account_card_name"
-        description="connect_link_account_card_description"
+        description="connect_unauthorised_account_description"
         context={context && context.text}
         button={{
           appearance: 'primary',
-          text: 'connect_link_account_card',
+          text: 'connect_unauthorised_account_action',
           testId: 'connect-account',
         }}
         onClick={onAuthorise}
         testId={testId}
-      />
+      >
+        <FormattedMessage
+          {...messages.connect_unauthorised_account_description}
+          values={{ context: context?.text }}
+        />{' '}
+        <a
+          href={CONTENT_URL_SECURITY_AND_PERMISSIONS}
+          target="_blank"
+          data-testid={`${testId}-learn-more`}
+        >
+          <FormattedMessage
+            {...messages.learn_more_about_smart_links_security}
+          />
+        </a>
+      </EmbedCardUnresolvedView>
     </ExpandedFrame>
   );
 };